backend/requirements.txt
langchain==0.3.19
fastapi==0.118.0
uvicorn==0.37.0
openai==1.61.1
langchain-community==0.3.17
langchain_openai==0.3.6
azure-core==1.35.1         
azure-functions==1.23.0                 
azure-identity==1.25.0             
azure-storage-blob==12.26.0    
azure-search-documents==11.5.3
python-dotenv
sse-starlette
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation==0.45b0
azure-monitor-opentelemetry-exporter
redis==7.1.0